Use strong type for Network.
The resulting type from parsing magic number was previously stringly
type. This is not considered idiomatic, as it comes with bunch of
problems: possible typos, decreased effiiency of comparison, lack of
static analysis...

This change introduces a new type called `Network` that implements a
bunch of standard traits for these kinds of types and this type is then
returned from parsing functions.