Added newtype for block time
This change adds a newtype for block time to improve robustness. Apart
from getting the advantages of strong type, this also removes the need
for storing the time in stringly form in the header, as it impls
Display. Finally, the Display implementation is more efficinet by not
allocating and clearer by using infallible conversion instead of
unwrapping.